requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
discord.py==1.7.3
dnspython==2.2.0
h11==0.12.0
idna==3.3
multidict==6.0.2
outcome==1.1.0
pycparser==2.21
pymongo==4.0.1
pyOpenSSL==21.0.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
wsproto==1.0.0
yarl==1.7.2
